Adds Arrow length <-> std size/offset conversions functions with tests
This change does not however uses these functions in the rest of the library.

We introduce 2 configuration options to enable the user to optionally constrain arrow data size to signed 32bit values,
and optionally enable runtime checks of arrow data size values.
The checks only happen in the new conversion functions so will not be effective until used in the rest of the library.